+++ src/Attributes/Collections/Pluck.php
@@ -1,0 +1,33 @@
+<?php declare(strict_types=1);
+
+/**
+ * Copyright (C) Brian Faust
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Cline\Struct\Attributes\Collections;
+
+use Attribute;
+use Cline\Struct\Contracts\TransformsLaravelCollectionValueInterface;
+use Cline\Struct\Support\CreationContext;
+use Illuminate\Support\Collection;
+
+/**
+ * @author Brian Faust <brian@cline.sh>
+ * @psalm-immutable
+ */
+#[Attribute(Attribute::TARGET_PROPERTY)]
+final readonly class Pluck implements TransformsLaravelCollectionValueInterface
+{
+    public function __construct(
+        public string|array $value,
+        public string|array|null $key = null,
+    ) {}
+
+    public function transformCollection(Collection $items, ?CreationContext $context = null): Collection
+    {
+        return $items->pluck($this->value, $this->key);
+    }
+}